+++ services/preview/resourcegraph/mgmt/2018-09-01/resourcegraph/CHANGELOG.md
@@ -1,2 +1,21 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Struct Fields
+
+1. GraphQueryResource.ETag
+1. GraphQueryUpdateParameters.ETag
+1. Resource.ETag
+
+## Additive Changes
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. GraphQueryResource.Etag
+1. GraphQueryUpdateParameters.Etag
+1. Resource.Etag